Implemented Encapsulation
The scope of the project was growing larger so I figured it was time to break everything apart and encapsulate them. I broke the functionality down into various different classes and spread them across different files to make the code easier to scale and easier to read as it becomes more complex.