Reworked Action System
I reworked the way actions are selected. Before actions would be randomly selected based on time and location. Now actions are based on character needs. Each character has various needs (which will be expanded upon later) that need to be met. The new system determines which need is the most pressing and the character will choose an action based on that. Each action will restore a certain amount of whatever need to caters to. This creates character schedules that feel more in line with real schedules and arent just completely random.